refactor: mockData.ts → types.ts 이름 변경 및 샘플 데이터 제거
