+++ frontend/src/services/api.ts
@@ -9,7 +9,7 @@ import {
   type Attachment,
   type AnalysisLog,
   type RequirementItem,
-} from '../components/mockData';
+} from '../types';
 
 const BASE_URL = import.meta.env.VITE_API_BASE_URL ?? 'http://localhost:8000';
